Continue with change 116008. Change from INT_MAX to LLONG_MAX also on Windows. I can't test on Windows, so bear with me if it doesn't compile out of the box. Use the Windows equivalent of LLONG_MAX if it's not available. Note that the LLONG_MAX availability is probably compiler-dependent, not platform-dependent on Windows.
[git-p4: depot-paths = "//depot/qt/main/": change = 116011]

+++ src/tools/qfile_wce.cpp
@@ -196,7 +196,7 @@ bool QFile::open( int m, FILE *f )
     if ( (st.st_mode & QT_STAT_MASK) != QT_STAT_REG ) {
 	// non-seekable
 	setType( IO_Sequential );
-	length = INT_MAX;
+	length = LLONG_MAX;
     } else {
 	length = (int)st.st_size;
     }
@@ -225,7 +225,7 @@ bool QFile::open( int m, int f )
     if ( (st.st_mode & QT_STAT_MASK) != QT_STAT_REG ) {
 	// non-seekable
 	setType( IO_Sequential );
-	length = INT_MAX;
+	length = LLONG_MAX;
     } else {
 	length = (int)st.st_size;
     }
